Rename filestructure according to endpoint path
