+++ client/state/data-layer/wpcom/index.js
@@ -3,12 +3,12 @@
  */
 import { mergeHandlers } from 'state/action-watchers/utils';
 import accountRecovery from './account-recovery';
+import activityLog from './activity-log';
 import comments from './comments';
 import me from './me';
 import plans from './plans';
 import posts from './posts';
 import read from './read';
-import rewind from './rewind';
 import sites from './sites';
 import timezones from './timezones';
 import themeFilters from './theme-filters';
@@ -17,12 +17,12 @@ import login2fa from './login-2fa';
 
 export const handlers = mergeHandlers(
 	accountRecovery,
+	activityLog,
 	comments,
 	me,
 	plans,
 	posts,
 	read,
-	rewind,
 	sites,
 	timezones,
 	themeFilters,
